+++ src/Dependencies.lua
@@ -151,6 +151,8 @@ gSounds = {
     ['walking-song1'] = love.audio.newSource('assets/music&SFX/walking-song1.mp3', 'static'),
     ['walking-song2'] = love.audio.newSource('assets/music&SFX/walking-song2.mp3', 'static'),
     ['stove-sound'] = love.audio.newSource('assets/music&SFX/stoveSound.mp3', 'static'),
+    ['stove'] = love.audio.newSource('assets/music&SFX/stove.mp3', 'static'),
+    ['chop'] = love.audio.newSource('assets/music&SFX/chop.mp3', 'static'),
 }
 
 require('src.states.BaseState')